requirements.txt
python-dotenv
azure-core
azure-identity
azure-search-documents
docx2txt
pypdf
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
notebook
